+++ src/routes/equipementGroup/group/listGroup.ts
@@ -101,7 +101,8 @@ module.exports = function (logger, app: express.Express, spinalAPIMiddleware: IS
           staticId: realNode.getId().get(),
           name: realNode.getName().get(),
           type: realNode.getType().get(),
-          color: group.color.get()
+          color: group.color.get(),
+          icon: group.icon?.get()
         };
         nodes.push(info);
       }