+++ src/routes/routes.ts
@@ -433,6 +433,11 @@ function routes(logger, app: express.Application, spinalAPIMiddleware: ISpinalAP
     app,
     spinalAPIMiddleware
   );
+  require('./geographicContext/building/buildingReferenceObjectsList')(
+    logger,
+    app,
+    spinalAPIMiddleware
+  );
   require('./geographicContext/floor/floorList')(
     logger,
     app,